feat(ui): implement Binance-style order book with price grouping
Refactor OrderBookHeatMap to match Binance's order book UI:

- Add dynamic price grouping dropdown with symbol-specific options
  (e.g., BTC shows 0.01, 0.1, 1, 10, 50, 100, 1000)
- Implement actual price aggregation logic that groups orders by
  price buckets and sums quantities
- Fix depth bar direction: bars now grow OUTWARD from center
  (where bid/ask prices meet)
- Color prices like Binance: green (#0ECB81) for bids,
  red (#F6465D) for asks
- Fix integer overflow in formatAsCurrency by using toLong()
  instead of toInt() for large price values
- Clean up unused imports and parameters

+++ composeApp/src/commonMain/kotlin/ktx/DoubleKtx.kt
@@ -5,7 +5,7 @@ import kotlin.math.roundToInt
 
 fun Double.formatAsCurrency(): String {
     val absValue = this.absoluteValue
-    val integerPart = absValue.toInt()
+    val integerPart = absValue.toLong()
     val fractionalPart = ((absValue - integerPart) * 100).roundToInt()
 
     val formattedInteger = integerPart.toString().reversed().chunked(3).joinToString(",").reversed()
